Set interwiki width to 18em all the time
The interwiki width was set to 17em by default, and 18em at screen sizes
greater than mobile. The width of the sidebar is 17em. The extra em was
to give the iframe a little extra internal space (in this case
theoretically half an em, but actually 5px) to display its box-shadow
without being eclipsed by the border of the iframe.

There is no reason, however, for the interwiki to be 17em on mobile. It
just looks weird. 18em is fine, so I have removed the media query and
made 18em the default.

The next commit will address the 5px offset, which should actually be
0.5em.