Increase interwiki offset from 5px to 0.5em
The interwiki is 1em wider than the sidebar to give it space to show its
shadow inside the iframe. It should be offset to the left by, and have
an internal margin of, half of that amount to centre-align it. However,
the offset is currently 5px, which is an unrelated value, presumably
originally decided by eye.

0.5em is the correct value.